+++ lib/main_dev.dart
@@ -8,8 +8,8 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ya_todo_app/config/themes/app_themes.dart';
 import 'package:ya_todo_app/core/data/api_client/api_client.dart';
-import 'package:ya_todo_app/core/data/api_client/queued_interceptor.dart';
-import 'package:ya_todo_app/core/data/api_client/revision_interceptor.dart';
+import 'package:ya_todo_app/core/data/api_client/api_interceptors/queued_interceptor.dart';
+import 'package:ya_todo_app/core/data/api_client/api_interceptors/revision_interceptor.dart';
 import 'package:ya_todo_app/core/data/local_data_source/hive_data_source.dart';
 import 'package:ya_todo_app/core/data/repository/revision_repository.dart';
 import 'package:ya_todo_app/core/di/di_container.dart';
