+++ lib/navigation/router_delegate.dart
@@ -1,7 +1,5 @@
 // ignore_for_file: public_member_api_docs
 
-import 'dart:developer';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:ya_todo_app/core/domain/serivces/app_loger.dart';
